Products controller and tests refactor
